Update readme
I added a section on installation and rewrote the authentication section
a bit.

I noticed that oclif has a way to automatically update the readme, so
I've used that to generate the readme. I hope you like it. One benefit
is that we don't have to update documentation in multiple places.